requirements.txt
Django>=3.2.0
djangorestframework>=3.12.0
django-cors-headers>=3.7.0
psycopg2-binary>=2.9.0
gunicorn>=20.1.0
whitenoise>=5.2.0
dj-database-url>=0.5.0
python-dotenv>=0.19.0
asgiref>=3.3.0
pytz>=2021.1
sqlparse>=0.4.0
